test: add registerNewUser helper that retries the submit under flake
The staging registration endpoint intermittently fails to return the
success confirmation under parallel-shard load — the recurring e2e flake
(profile.spec failed here, unrelated to the add-to-list fix). Add a
shared registerNewUser helper that waits out the hydration gate and
retries the submit once, and use it in profile and add-to-list specs.
